+++ tests/0200_factory/0201_card.py
@@ -17,10 +17,19 @@ def test_factory(db):
 
 def test_feature_getters(db):
     """
-    Features getters should return the right features with scope and plugins enforced.
+    Features getters should return the right features with scope and plugins enforced,
+    also the scopes values never contain duplicate classnames.
+
+    This does not care about setting 'OOO' because it is not used in getters, they
+    correclyt manage value with or without whitespace.
     """
-    size_foo = FeatureFactory(title="Foo", scope="size", plugins=["Card"])
-    size_bar = FeatureFactory(title="Bar", scope="size", plugins=[
+    size_foo = FeatureFactory(
+        title="Foo",
+        value="foo bis",
+        scope="size",
+        plugins=["Card"]
+    )
+    size_bar = FeatureFactory(title="Bar", value="bis bar", scope="size", plugins=[
         "Album",
         "Card"
     ])
@@ -32,6 +41,7 @@ def test_feature_getters(db):
     ])
     extra_bang = FeatureFactory(title="bang", scope="extra", plugins=["Card"])
 
+    # A card defining some feature well assigned to their scope, everything is well
     first_card = CardFactory(
         fill_size_features=[size_foo, size_bar],
         fill_color_features=[color_ping, color_foo, color_zap],
@@ -40,10 +50,12 @@ def test_feature_getters(db):
     assert first_card.scoped_features() == {
         "color": ["foo", "ping"],
         "extra": [],
-        "size": ["bar", "foo"],
+        "size": ["bar", "bis", "foo"],
     }
-    assert first_card.flat_features() == "bar foo ping"
+    assert first_card.flat_features() == "bar bis foo ping"
 
+    # A card defining some features in the wrong scopes, those wrongly scoped features
+    # will be ignored
     second_card = CardFactory(
         fill_size_features=[size_foo, color_foo],
         fill_color_features=[color_ping, color_foo, color_zap],
@@ -53,6 +65,6 @@ def test_feature_getters(db):
     assert second_card.scoped_features() == {
         "color": ["foo", "ping"],
         "extra": ["bang"],
-        "size": ["foo"],
+        "size": ["bis", "foo"],
     }
-    assert second_card.flat_features() == "bang foo ping"
+    assert second_card.flat_features() == "bang bis foo ping"