+++ tests/0400_admin/0400_feature.py
@@ -33,10 +33,16 @@ def test_export_empty(db, admin_client):
 def test_export_content(db, admin_client, settings):
     """
     Exporting view should work without any error when there is some features.
+
+    Also note how the exporter does not care about value syntax (Ping item contains
+    whitespaces even it is not allowed from settings) because it is assumed that stored
+    content has been well validated.
     """
+    settings.BLOCKS_FEATURE_ALLOW_MULTIPLE_CLASSES = False
+
     FeatureFactory(title="Foo", scope="size", plugins=["Card"])
     FeatureFactory(title="Bar", scope="size", plugins=["Album,Card"])
-    FeatureFactory(title="Ping", scope="color", plugins=["Hero"])
+    FeatureFactory(title="Ping", value="ping floyd", scope="color", plugins=["Hero"])
     FeatureFactory(title="Pong", scope="extra", plugins=["Hero"])
 
     url = reverse("admin:cmsplugin_blocks_feature_export")
@@ -47,7 +53,7 @@ def test_export_content(db, admin_client, settings):
     assert payload["items"] == [
         {
             "title": "Ping",
-            "value": "ping",
+            "value": "ping floyd",
             "scope": "color",
             "plugins": [
                 "Hero"